% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crest}
\alias{crest}
\title{crest}
\format{
An object of class \code{list} of length 8.
}
\usage{
crest
}
\description{
Murine neural crest data,doi: 10.1126/science.aas9536
}
\examples{
  head(crest)
}
\keyword{datasets}
