% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateDE.R
\name{Determine_transition_state}
\alias{Determine_transition_state}
\title{Identify index of pre somatic cell cells}
\usage{
Determine_transition_state(exprs_mat)
}
\arguments{
\item{exprs_mat}{Gene expression matrix, cells by now, genes by column}
}
\value{
Index of pre somatic cell cells
}
\description{
Identify index of pre somatic cell cells
}
