% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateDE.R
\name{fateDE_fit_one}
\alias{fateDE_fit_one}
\title{Fitting gene expression of one branch}
\usage{
fateDE_fit_one(df1_notfit)
}
\arguments{
\item{df1_notfit}{Branch 1 unfitted data}
}
\value{
a dataframe of fitting
}
\description{
Fitting gene expression of one branch
}
