% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FateDE.R
\name{FateDE_main}
\alias{FateDE_main}
\title{FateDE: Searching for Destiny Determinants}
\usage{
FateDE_main(
  branch1,
  branch2,
  G1 = NULL,
  G2 = NULL,
  GP_FSV = 0.9,
  logN = T,
  TFs = NULL,
  cluster_meth = "ward.D2"
)
}
\arguments{
\item{branch1}{A Dataframe,cells by row, the first column is Pseudotime, the second column is cell State,
other columns are genes.}

\item{branch2}{The same as branch1.}

\item{G1}{A vector,the set of genes that increase in expression over pseudotime in branch 1.}

\item{G2}{A vector,the set of genes that increase in expression over pseudotime in branch 2.}

\item{GP_FSV}{A number, screening threshold in Gaussian process.}

\item{logN}{A logic value, is a logified expression matrix used in Gaussian processes.}

\item{TFs}{A vector, extract the TF of the expression matrix for calculation.}

\item{cluster_meth}{clustering method, clustering methods using the same as pheatmap.}
}
\value{
A list, include G1, G2, GP's results, G1_1, G2_1, one_two_cor_pairwise_all's results
}
\description{
FateDE: Searching for Destiny Determinants
}
