% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hepatoblast}
\alias{hepatoblast}
\title{hepatoblast}
\format{
An object of class \code{dynwrap::with_cell_waypoints} (inherits from \code{dynwrap::with_prior}, \code{dynwrap::with_expression}, \code{dynwrap::with_trajectory}, \code{dynwrap::with_grouping}, \code{dynwrap::data_wrapper}, \code{list}) of length 21.
}
\usage{
hepatoblast
}
\description{
Hepatoblast differentiation data,doi: 10.1002/hep.29353
}
\examples{
  head(hepatoblast)
}
\keyword{datasets}
