% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_DoubleGenes_trend}
\alias{plot_DoubleGenes_trend}
\title{Draw expression curves for two genes.}
\usage{
plot_DoubleGenes_trend(
  df_prepare,
  df1,
  df2,
  double_gene,
  color_by = "State",
  cell_size = 1.5,
  min_expr = NULL
)
}
\arguments{
\item{df_prepare}{A list, unfitted gene expression data for two branches.}

\item{df1}{A dataframe, gene expression data for fate 1.}

\item{df2}{A dataframe, gene expression data for fate 2.}

\item{double_gene}{Gene vector}

\item{color_by}{Color by}

\item{cell_size}{cell size}

\item{min_expr}{Min expr}
}
\value{
ggplot
}
\description{
Draw expression curves for two genes.
}
