% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Genes_Heatmap}
\alias{plot_Genes_Heatmap}
\title{Draw a heat map of gene expression.}
\usage{
plot_Genes_Heatmap(
  df1,
  df2,
  G1_1,
  branch_labels = c("fate1", "fate2"),
  branch_colors = c("#979797", "#F05662", "#7990C8"),
  use_gene_short_name = TRUE,
  show_rownames = T,
  hclust_method = "ward.D2",
  num_clusters = 6,
  scale_max = 3,
  scale_min = -3
)
}
\arguments{
\item{df1}{A dataframe, gene expression data for fate 1.}

\item{df2}{A dataframe, gene expression data for fate 2.}

\item{G1_1}{Gene vector.}

\item{branch_labels}{A vector.}

\item{branch_colors}{A vector}

\item{use_gene_short_name}{T/F}

\item{show_rownames}{T/F}

\item{hclust_method}{hclust method}

\item{num_clusters}{number of clusters}

\item{scale_max}{scale max}

\item{scale_min}{scale min}
}
\value{
ggplot
}
\description{
Draw a heat map of gene expression.
}
