% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_SingleGene_trend}
\alias{plot_SingleGene_trend}
\title{Plot a curve of a single gene over pseudotime on different branches.}
\usage{
plot_SingleGene_trend(
  df_prepare,
  df1,
  df2,
  gene,
  color_by = "State",
  cell_size = 1.5,
  min_expr = NULL
)
}
\arguments{
\item{df_prepare}{A list, unfitted gene expression data for two branches.}

\item{df1}{A dataframe, gene expression data for fate 1.}

\item{df2}{A dataframe, gene expression data for fate 2.}

\item{gene}{Gene name.}

\item{color_by}{Color by}

\item{cell_size}{Cell size}

\item{min_expr}{Min expr}
}
\value{
ggplot
}
\description{
Plot a curve of a single gene over pseudotime on different branches.
}
