% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{simulation}
\format{
An object of class \code{array} of dimension 12 x 1001 x 200.
}
\usage{
simulation
}
\description{
Central nervous system simulation data generated by stochastic differential equation
}
\examples{
  head(simulation)
}
\keyword{datasets}
